SSCS-10371-create-dynamic-list-override : handover commit
